tercer intento

+++ download-plugins.sh
@@ -1,15 +1,16 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -e
 
 PLUGIN_DIR="/data/plugins"
-
 mkdir -p "$PLUGIN_DIR"
 
-URL=$(curl -s https://api.modrinth.com/v2/project/heartbound/version \
-    | jq -r '.[0].files[0].url')
+if [ ! -f "$PLUGIN_DIR/Heartbound.jar" ]; then
+    echo "Descargando Heartbound..."
 
-curl -L "$URL" -o "$PLUGIN_DIR/Heartbound.jar"
+    URL=$(curl -s https://api.modrinth.com/v2/project/heartbound/version \
+        | jq -r '.[0].files[0].url')
 
-chown -R minecraft:minecraft /data
+    curl -L "$URL" -o "$PLUGIN_DIR/Heartbound.jar"
 
-exec su-exec minecraft /start
+    chown -R 1000:1000 "$PLUGIN_DIR"
+fi
